+++ src/main/java/org/folio/persist/HoldingsRepository.java
@@ -2,11 +2,13 @@ package org.folio.persist;
 
 import static org.folio.rest.impl.HoldingsStorageApi.HOLDINGS_RECORD_TABLE;
 import static org.folio.rest.persist.PgUtil.postgresClient;
+import static org.folio.services.location.LocationService.LOCATION_TABLE;
 
 import io.vertx.core.Context;
 import io.vertx.core.Future;
 import io.vertx.sqlclient.Row;
 import io.vertx.sqlclient.RowSet;
+import io.vertx.sqlclient.Tuple;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
@@ -17,6 +19,53 @@ import org.folio.rest.persist.cql.CQLWrapper;
 public class HoldingsRepository extends AbstractRepository<HoldingsRecord> {
   public HoldingsRepository(Context context, Map<String, String> okapiHeaders) {
     super(postgresClient(context, okapiHeaders), HOLDINGS_RECORD_TABLE, HoldingsRecord.class);
+  }
+
+  /**
+   * Row where the first value is an array with all holdings records; the second value is
+   * the exact totalRecords count.
+   */
+  public Future<Row> getByInstanceId(String instanceId, String[] sortBys, int offset, int limit) {
+    var orderBy = new StringBuilder();
+    for (var sortBy : sortBys) {
+      if (sortBy.isEmpty()) {
+        continue;
+      }
+      var s = switch (sortBy) {
+        case "effectiveLocation.name" -> "name";
+        case "callNumberPrefix",
+             "callNumber",
+             "callNumberSuffix" -> "jsonb->>'" + sortBy + "'";
+        default -> null;
+      };
+      if (s == null) {
+        return Future.failedFuture(new IllegalArgumentException("sortBy: " + sortBy));
+      }
+      if (!orderBy.isEmpty()) {
+        orderBy.append(", ");
+      }
+      orderBy.append(s);
+    }
+    var sql = "WITH data AS ("
+        + " SELECT h.jsonb AS jsonb, l.jsonb->>'name' AS name"
+        + " FROM " + HOLDINGS_RECORD_TABLE + " h"
+        + " LEFT JOIN " + LOCATION_TABLE + " l"
+        + "   ON h.effectiveLocationId = l.id"
+        + " WHERE h.instanceId=$1"
+        + " )"
+        + " SELECT json_array("
+        + "   SELECT jsonb"
+        + "   FROM data"
+        + "   ORDER BY " + orderBy
+        + "   OFFSET $2"
+        + "   LIMIT $3"
+        + " )::text, ("
+        + "   SELECT COUNT(*)"
+        + "   FROM " + HOLDINGS_RECORD_TABLE
+        + "   WHERE instanceId=$1"
+        + ")";
+    return postgresClient.withReadConn(conn -> conn.execute(sql, Tuple.of(instanceId, offset, limit)))
+        .map(rowSet -> rowSet.iterator().next());
   }
 
   /**